Split store over multiple files
